requirements.txt
certifi==2024.6.2
chardet==5.2.0
idna==3.7
pkginfo==1.11.1
requests==2.32.3
requests-toolbelt==1.0.0
tqdm==4.66.4
twine==5.1.1
urllib3==2.2.2
split-file-reader==0.1.4
pyyaml==6.0.1
platformdirs==4.2.2
py-robocopy==0.0.5
